Make SplObjectStorage exporting in HHVM match Zend
SplObjectStorage::$storage is the current method.
HHVM is about to be changed so that it uses $__storage instead to reduce
the chance of a conflict in cases like this:

```PHP
<?php

$o1 = new stdClass();
$store = new SplObjectStorage;
$store->attach($o1);
$store->storage = 'herpderp';
```

This gets PHPUnit to 100%, but #1 still depends on an array_keys fix for
recursive structures that hasn't reached master yet.