Improve export of closed resources (for 3.x)
As the 3.x branch has a minimum PHP version of PHP 7.0 and the `resource (closed)` return value for `gettype()` was only introduced in PHP 7.2, a slightly more involved solution for the same is needed for the 3.x branch.

The methods I'm adding are based on code I've used to polyfill the `assertClosedResource()` assertion in the PHPUnit Polyfills and has been (and still is) extensively tested there.

Fixes 36